add details to tag_info packet
I have a session file when there were three tags connected to the base, and it was pretty clear what two of these numbers are. tag_registration_id is the number displayed on the tag screen. E.g. I have one that says T5_foobarbecue -- it's the fifth tag I've registered to that username. The tag_connection_id seems to correspond to the tag id in the GPS packets, and I think to the order of appearance in the Tag Manager screen on the base.